+++ u-sdk/tests/oss/main.rs
@@ -205,14 +205,27 @@ fn put_object_presigned_url_test() {
 fn generate_post_object_policy_test() {
     let now = OffsetDateTime::now_utc();
     let client = get_oss_client();
+
+    let callback_body = CallBackBody::builder()
+        .bucket(true)
+        .object(true)
+        .vars([("uid", "uid", "1234"), ("order", "order_id", "1234")])
+        .build();
+    let callback = OssCallBack::builder()
+        .callback_url(["https://example.com/webhook/oss/callback"])
+        .callback_body(callback_body)
+        .callback_sni(false)
+        .callback_body_type(CallbackBodyType::Json)
+        .build();
+
     let res = client
         .post_object()
         .content_length_range((1, 1024 * 1024))
-        .key(("eq".to_owned(), "test2/t-sample.txt".to_owned()))
-        .x_oss_forbid_overwrite("true".to_owned())
+        .key(("eq".to_owned(), "test2/t-sample7777.txt".to_owned()))
         .x_oss_content_type("text/plain".to_owned())
         .x_meta("kk1", ("eq", "vv1"))
         .x_metas([("kk2", ("eq", "vv2")), ("kk3", ("starts-with", "vv3"))])
+        .callback(callback)
         .build()
         .generate_policy(now + Duration::from_secs(900));
     println!("res: {:#?}", res);